Feat & Fix: Make the restart button functional and fix bug causing the game over dialog to show before the game is even played.
I made the restart button functional. It resets everything concerning
that round, such as the points earned by the computer and the user, as well
as the level or number of rounds that was being played. From the last update
I merged  3 conditional statements into one and as such, it created a bug
that made the "game over" modal show before the game is even played. So I seperated
one, the one for the custom level and the bug was fixed. I'm yet to find out why.